refactor(store): one schema owner — search SQL and serve queries move into store/, scoring inverts out
- serve/data.rs no longer runs raw sqlx against the store schema: all five
  builders' query bodies move into store/serve_queries.rs as typed-row methods;
  serve keeps only wire-shaping (injection flags, trust level, preview
  truncation, NodeRef/Node/Edge construction, BFS/dedup/degree). Zero sqlx in
  serve.
- RRF fusion (rrf_fuse, rrf_fuse_n, rrf_k, set_rrf_k_from_config) moves out of
  store/search.rs into search/scoring/fusion.rs as free functions — search owns
  fusion. The ASCII name-match helpers move to store/helpers/scoring.rs next to
  score_name_match_pre_lower.
- store/search.rs now imports nothing from crate::search, killing the
  knob / BoundedScoreHeap / rrf_fuse back-edges. It retains only store-owned SQL
  (search_fts, fts_match_ids, search_by_name) with the v27 needs_embedding gate
  intact.
- Pure code motion: bodies move verbatim, logic unchanged. v27 visibility gates,
  fused-score reuse path, and note-boost cell caching survive verbatim.
