+++ src/search/scoring/mod.rs
@@ -11,6 +11,7 @@
 mod candidate;
 mod config;
 mod filter;
+mod fusion;
 pub mod knob;
 mod name_match;
 mod note_boost;
@@ -19,5 +20,7 @@ pub(crate) use candidate::{
     apply_parent_boost, apply_scoring_pipeline, score_candidate, BoundedScoreHeap, ScoringContext,
 };
 pub(crate) use filter::{build_filter_sql, compile_glob_filter, extract_file_from_chunk_id};
+pub(crate) use fusion::rrf_fuse;
+pub use fusion::set_rrf_k_from_config;
 pub(crate) use name_match::NameMatcher;
 pub(crate) use note_boost::{NoteBoost, NoteBoostIndex, OwnedNoteBoostIndex};
